+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 
 ## 0.4.0 (unreleased)
 
+- Generalized name of model ``Website`` to ``Resource``.
+
 
 ## 0.3.0 (2024-05-16)